move lexical tests to its own suite. pnpm dev fields is already compiling, pnpm dev lexical not yet
