+++ test/lexical/collections/Lexical/blockComponents/LabelComponent.tsx
