+++ test/lexical/collections/Lexical/inlineBlockComponents/LabelComponent.tsx
