+++ test/lexical/collections/LexicalAccessControl/index.ts
