+++ test/lexical/collections/LexicalLocalized/generateLexicalRichText.ts
