+++ test/lexical/collections/LexicalLocalized/textToLexicalJSON.ts
