+++ test/lexical/collections/LexicalObjectReferenceBug/index.ts
