+++ test/lexical/collections/RichText/data.ts
