+++ test/lexical/collections/RichText/e2e.spec.ts
