+++ test/lexical/collections/RichText/generateLexicalRichText.ts
