+++ test/lexical/collections/RichText/generateSlateRichText.ts
