+++ test/lexical/collections/RichText/index.ts
