+++ test/lexical/lexical.int.spec.ts
@@ -13,9 +13,17 @@ import type { PaginatedDocs, Payload } from 'payload'
 import path from 'path'
 import { fileURLToPath } from 'url'
 
-import type { LexicalField, LexicalMigrateField, RichTextField } from './payload-types.js'
+import type { LexicalField, LexicalMigrateField, RichTextField } from '../fields/payload-types.js'
 
 import { devUser } from '../credentials.js'
+import {
+  arrayFieldsSlug,
+  lexicalFieldsSlug,
+  lexicalMigrateFieldsSlug,
+  richTextFieldsSlug,
+  textFieldsSlug,
+  uploadsSlug,
+} from '../fields/slugs.js'
 import { initPayloadInt } from '../helpers/initPayloadInt.js'
 import { NextRESTClient } from '../helpers/NextRESTClient.js'
 import { lexicalDocData } from './collections/Lexical/data.js'
@@ -27,14 +35,6 @@ import { generateLexicalRichText } from './collections/RichText/generateLexicalR
 import { textDoc } from './collections/Text/shared.js'
 import { uploadsDoc } from './collections/Upload/shared.js'
 import { clearAndSeedEverything } from './seed.js'
-import {
-  arrayFieldsSlug,
-  lexicalFieldsSlug,
-  lexicalMigrateFieldsSlug,
-  richTextFieldsSlug,
-  textFieldsSlug,
-  uploadsSlug,
-} from './slugs.js'
 
 let payload: Payload
 let restClient: NextRESTClient
