+++ test/lexical/slugs.ts
@@ -1,0 +1,6 @@
+export const lexicalFieldsSlug = 'lexical-fields'
+export const lexicalLocalizedFieldsSlug = 'lexical-localized-fields'
+export const lexicalMigrateFieldsSlug = 'lexical-migrate-fields'
+export const lexicalRelationshipFieldsSlug = 'lexical-relationship-fields'
+export const lexicalAccessControlSlug = 'lexical-access-control'
+export const richTextFieldsSlug = 'rich-text-fields'
